+++ fuzz/crl.c
@@ -26,7 +26,7 @@ int FuzzerTestOneInput(const uint8_t *buf, size_t len)
     const unsigned char *p = buf;
     unsigned char *der = NULL;
 
-    X509_CRL *crl = d2i_X509_CRL(NULL, &p, len);
+    X509_CRL *crl = d2i_X509_CRL(NULL, &p, (long)len);
     if (crl != NULL) {
         BIO *bio = BIO_new(BIO_s_null());
         X509_CRL_print(bio, crl);
